+++ zcode/inout/__init__.py
@@ -5,7 +5,10 @@ from . import inout_core
 from .inout_core import *
 from . import log
 from .log import *
+from . import timer
+from .timer import *
 
 __all__ = []
 __all__.extend(inout_core.__all__)
 __all__.extend(log.__all__)
+__all__.extend(timer.__all__)